WhenThereAreExceptionsInMessageProcessing
